test: adjust test preps for BalanceProvider test: adjust test preps for BalanceUpdater
